[NVMESH-8062] [[DbitsDeg] part3, Bugfix: need commit binfo if ow!=rldr
Due to merge of multiple binfo from all locks, pre can differ
from locks. Pre can have 0 dbit while each binfo can be != 0
We need to syncronize binfo copies. if binfo of rldr is not equal
then we need to syncronize to this (deduced) value, not to value
of primary owner lock

Change-Id: I7b840cfee1b1206e0078797ed465ad58e28a1798
Signed-off-by: danielhsh <DanielHe@nvidia.com>
Reviewed-by: Daniel Gan-Levi <dganlevi@nvidia.com>