[NVMESH-6934] init-scripts: sense and use a client upgrade directive
Take into account a one-time file /var/run/nvmesh/nvmeshclient/upgrade
and set a variable marking we are within an upgrade (NDU).

For stop flow, revive the flow that sets the upgrade directive for the shutdown script.

For start flow, delete the file when succes.

The file would be created by the upgrade agent.

Change-Id: I4c044f4cc2e16ab2bb64e546667c53088ac8e5a7
Signed-off-by: Or Gerlitz <ogerlitz@nvidia.com>