[NVMESH-7022]: NRCH: Fix error trace in iommu mode
In new iommu mode where we wait for both completions,
reuse_bb was already reset when nvmeibc_ib_net_complete_iocmd_sg was
called and so was printing an error trace for every reuse write.

This was solved by returning was_reuse from del_use_request
and pass it to nvmeibc_ib_net_complete_iocmd_sg.

This required adding was_reuse as a parameter to
- nvmeibc_ib_net_complete_iocmd_sg
- nvmeibc_ib_net_complete_iocmd
- nvmeibc_ib_net_unmap_and_unlink_iocmd

To keep code-changes small, a new version of these
functions was created with the was_reuse parameter
and macros were used to redirect the calls where
it is not relevant to the new version with was_reuse=false

Change-Id: Idcce59d820f1110943c336e2fcc39cff338f559b
Signed-off-by: Jared <jholzman@nvidia.com>