[NVMESH-6936] client NDU breakdown tool - add offline analysis
Added logs-dir and pager optional params to the analyze mode

The logs-dir structure is assumed to from the log-collector, e.g. containing copy of
var/log/nvmesh/trace_daemon/[..] and journalctl_curr/prev_boot_log_-7200_seconds files.

The pager params comes to serve a case where the pager executable from the collected logs
doesn't run due to python or libc issues.

The PagerLog class was enhanced to work with offline logs.

A JournalctlFileLog class is introduced, it filters lines from the file and for those
of interest (match the identifiers we care about and in the time range) it creates
json record which is sub-set of the full record returned by journalctl on non-offline analysis.

For clarity, this is what we are emulating:

journalctl -o json --since 2025-11-23 09:28:12.000000 --until 2025-11-23 09:33:12.000000 \
		SYSLOG_IDENTIFIER=systemd SYSLOG_IDENTIFIER=nvmeshclient

Change-Id: I8a577f531752dd632d974f0f7fb05af55db7b645
Signed-off-by: Or Gerlitz <ogerlitz@nvidia.com>