[NVMESH-7253] Block: Fix wrong (non-NULL) holder parameter to bdev[_file]_open_by_path()
The semantics of this parameter have changed from the previous
incarnation of this function (blkdev_get_by_path): now it is considered
also if the requested mode is not exclusive, so if there is already an
exclusive open on the bdev, the open will always fail.
We used to pass a non-NULL holder parameter apparently so that exclusive opens
would be also supported in __get_safe_kern_dev(), but we never open in
exclusive mode with this function, so this is unnecessary.

Change-Id: I3c6042c58ed466d23b5c26631c155118d9a708fc
Signed-off-by: Daniel Gan-Levi <dganlevi@nvidia.com>