[NVMESH-6936] tools/client NDU breakdown: move validation and error classification into the parser
Make get_entries (the iterator) to be a dump iterator, no longer decided on exceptions.

Make _parse_line (the parser) to be responsible for validation and error classification.

We don't expect any empty traces for NDU related ones, as such we check/act on
empty messages only once and in _parse_line.

Change-Id: I4ec57ed1eaab4e4f517c8b10e9aef14e7cc6660b
Signed-off-by: Or Gerlitz <ogerlitz@nvidia.com>