TomaSimu: add new TSB_socket_find_by_fd_opt() for fallible lookup
In order to lookup sandbox fd's and handle the case where the fd is not in the
list, without crashing, add an alternative function TSB_socket_find_by_fd_opt()
which returns an "optional" TSB pointer, where the original function guarantees
it always returns a TSB pointer or it aborts() if it can't.

This avoids a crash in close() when Toma is doing a graceful shutdown and
closing all fd's, in particular there was one for /dev/urandom that was
crashing it.

Change-Id: I68056a022539dee26e9c1c1067efc970b6dc50ee
Signed-off-by: Colin Bennett <cbennett@nvidia.com>